+++ site/China/bjdpm/bjdpm.go
@@ -70,6 +70,9 @@ func Download(dt *DownloadTask) (msg string, err error) {
 	storePath := dt.SavePath + string(os.PathSeparator)
 	ext := "." + dziFormat.Format
 	outfile := storePath + dt.BookId + ext
+	if util.FileExist(dest) {
+		return "", nil
+	}
 	if ret := util.StartProcess(dest, outfile, args); ret == true {
 		os.Remove(dest)
 	}
